refactor: view toggle
